Refined Level Manager and Finished Melee Attack
After a lot of back and forth trying to manage data persistency between levels as well as ensuring the game can still run even if mistakes in level are present, I finally have a good working solution that will allow the game to play if the level design is broken while still maintaining the rest of the data that isn't broken for player and player indicator original placement in level. I also have a working melee attack for the player. It is a basic melee attack that is better suited for a parry than actual melee, but game designers can set it up however they want.