Added weapon type to data persistency, fixed character changer
The current weapon now persists between scenes and saves which was an oversight I made when writing this solution. I also fixed issues with swapping characters with the press of a button. The only downside is that it the characters always spawn facing right; there is a fix somewhere for this, but the foundation on how the character should be facing in the scene all ties back to the very start of the scene and would require rewriting significant parts of the solution to have character facing the similar direction. It also adds a weapon pool to the scene even if that pool already exists.