Added a check to the PlayerHealth
In case the CharacterManager episode hasn't been watched or the feature hasn't been added, there is  a check to see if the character has the CharacterManager script on it then it checks for logic when player dying. This will prevent an error for those that don't want the CharacterManager but still have the refactored script.